Move `ScmIntegrationIcon`
