+++ .changeset/long-socks-fold.md
@@ -1,0 +1,8 @@
+---
+'@backstage/integration-react': patch
+'@backstage/plugin-catalog': patch
+---
+
+Move `ScmIntegrationIcon` from `@backstage/plugin-catalog` to
+`@backstage/integration-react` and make it customizable using
+`app.getSystemIcon()`.
