+++ packages/integration-react/src/api/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Spotify AB
+ * Copyright 2021 Spotify AB
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
-export { AboutCard } from './AboutCard';
-export { AboutContent } from './AboutContent';
-export { AboutField } from './AboutField';
+export {
+  ScmIntegrationsApi,
+  scmIntegrationsApiRef,
+} from './ScmIntegrationsApi';
